Correct usage of ttest_ind in Lab 5
In the last correction to this lab, I was too conservative about
changing this problem, and left in a line ostensibly taking a t-test
between a set of samples and a single sample. This corrects that part of
the lab technically, but is a bit clunky. It may be best to simply
exclude this part of the jury selection question.